CSS edited
I have a couple of problems…
1) when clicking NAV on header, for example, On clicking About , it
moves to “About” area but the text “about” is hidden. Same things to
the rest of NAV.
2) I would like “Twitter, Facebook Logo” to be color “white” before
clicking but it shows up as Purple..

3) I would like “main image” to show with Fade in / out not sliding
sideways. I researched how to do it.. and it looks like it can be
controlled in CSS not JS. Correct????